Add support for some block devices in /dev under Linux
There appears to be no generic cross-platform way to determine the size
of a block device, but if we’re running on Linux, we can discover the
size of some of them from sysfs(5). This assumes that devices are placed
under their normal names under /dev (no subdirectories, since we need
the normal basename for the sysfs path) and that sysfs is mounted under
its usual location.

The format of /sys/block/*/size does not seem to be documented anywhere,
but the size is always in 512-byte sectors (not device-specific blocks)
according to the comments on a Unix & Linux StackExchange answer [1].

Fixes #1.

[1]: https://unix.stackexchange.com/a/52219/44049