Don’t print trailing space if ETA is missing
Instead of printing the percentage and the ETA separated by a space,
even if the ETA is absent (undefined variable = empty string), prepend a
space to the ETA and concatenate them directly, so that there is no
trailing space if the ETA could not be determined yet.

Fixes #3, where the trailing space confused test-pmonitor.sh.